raidi

+++ admin/includes/view_all_posts.php
@@ -1,6 +1,48 @@
+<?php
+if (isset($_POST['checkBoxArray'])) {
+  foreach ($_POST['checkBoxArray'] as $postValueId) {
+    $bulk_options = $_POST['bulk_options'];
+
+      switch ($bulk_options) {
+        case 'published':
+          $query = "UPDATE posts SET post_status = '{$bulk_options}' WHERE post_id = {$postValueId} ";
+          $update_to_published_status = mysqli_query($connection, $query);
+          confirmQuery($update_to_published_status);
+          break;
+
+        case 'draft':
+          $query = "UPDATE posts SET post_status = '{$bulk_options}' WHERE post_id = {$postValueId} ";
+          $update_to_draft_status = mysqli_query($connection, $query);
+          confirmQuery($update_to_draft_status);
+          break;
+
+        case 'delete':
+          $query = "DELETE FROM posts WHERE post_id = {$postValueId} ";
+          $update_to_delete_status = mysqli_query($connection, $query);
+          confirmQuery($update_to_delete_status);
+          break;
+      }
+  }
+}
+ ?>
+<form action="" method="post">
 <table class="table table-bordered table-hover">
+  <div id="bulkOptionsContainer" class="col-xs-4">
+    <select class="form-control" name="bulk_options">
+      <option value="">Select Options</option>
+      <option value="published">Publish</option>
+      <option value="draft">Draft</option>
+      <option value="delete">Delete</option>
+    </select>
+  </div>
+  <div class="col-xs-4">
+    <input type="submit" name="submit" value="Apply" class="btn btn-success">
+    <a class="btn btn-primary" href="posts.php?source=add_post">Add New</a>
+  </div>
+  </div>
     <thead>
         <tr>
+            <th><input id="selectAllBoxes" type="checkbox"></th>
             <th>Id</th>
             <th>Author</th>
             <th>Title</th>
@@ -10,8 +52,10 @@
             <th>Tags</th>
             <th>Comments</th>
             <th>Date</th>
+            <th>View Post</th>
             <th>Edit</th>
             <th>Delete</th>
+            <th>Views</th>
         </tr>
     </thead>
     <tbody>
@@ -31,8 +75,14 @@
             $post_tags = $row['post_tags'];
             $post_comment_count = $row['post_comment_count'];
             $post_date = $row['post_date'];
+            $post_views_count = $row['post_views_count'];
 
             echo "<tr>";
+            ?>
+
+              <td><input class='checkBoxes' type='checkbox' name='checkBoxArray[]' value='<?php echo $post_id; ?>'></td>
+
+            <?php
                 echo "<td>{$post_id}</td>";
                 echo "<td>{$post_author}</td>";
                 echo "<td>{$post_title}</td>";
@@ -53,14 +103,17 @@
                 echo "<td>{$post_tags}</td>";
                 echo "<td>{$post_comment_count}</td>";
                 echo "<td>{$post_date}</td>";
+                echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
                 echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td><a href='posts.php?reset={$post_id}'>{$post_views_count}</a></td>";
             echo "</tr>";
         }
 
         ?>
     </tbody>
 </table>
+</form>
 
 <?php
 
@@ -70,5 +123,10 @@ if (isset($_GET['delete'])) {
     $delete_query = mysqli_query($connection, $query);
     header("Location: ./posts.php");
 }
-
+if (isset($_GET['reset'])) {
+    $the_post_id = $_GET['reset'];
+    $query = "UPDATE posts SET post_views_count = 0 WHERE post_id =" . mysqli_real_escape_string($connection, $_GET['reset']) . " ";
+    $reset_query = mysqli_query($connection, $query);
+    header("Location: ./posts.php");
+}
  ?>
